Add fail to image drivers
If creating an image failed, brickstrap would still complete with
a successful exit status. set -e does not have an effect here, so
checking the return status and failing with a message if any of the
image creation commands fail.

+++ brickstrap-image-drivers.sh
@@ -40,6 +40,7 @@ function brp_image_drv_single_fs()
         mount /dev/sda1 / : \
         tar-in "$(br_tarball_path)" / : \
 
+    [ "$?" -eq "0" ] || fail "Creating image failed"
 }
 
 #
@@ -102,19 +103,23 @@ function brp_image_drv_bootroot()
         mount /dev/sda1 /boot/flash : \
         tar-in "$(br_tarball_path)" / : \
 
+    [ "$?" -eq "0" ] || fail "Creating image failed"
+
     # Handle special case of BeagleBone boot loader
     # See http://elinux.org/Beagleboard:U-boot_partitioning_layout_2.0
     if [ -z "$BB_MLO_FILE" ] && [ -f "$(br_rootfs_dir)$(br_beaglebone_boot_dir)/MLO" ]; then
         BB_MLO_FILE="$(br_rootfs_dir)$(br_beaglebone_boot_dir)/MLO"
     fi
     if [ -f "$BB_MLO_FILE" ]; then
-        dd if="$BB_MLO_FILE}"of="$1" count=1 seek=1 bs=128k conv=notrunc
+        dd if="$BB_MLO_FILE" of="$1" count=1 seek=1 bs=128k conv=notrunc
+        [ "$?" -eq "0" ] || fail "Writing MLO failed"
     fi
     if [ -z "$BB_UBOOT_IMG_FILE" ] && [ -f "$(br_rootfs_dir)$(br_beaglebone_boot_dir)/u-boot.img" ]; then
         BB_UBOOT_IMG_FILE="$(br_rootfs_dir)$(br_beaglebone_boot_dir)/u-boot.img"
     fi
     if [ -f "$BB_UBOOT_IMG_FILE" ]; then
         dd if="$BB_UBOOT_IMG_FILE" of="$1" count=2 seek=1 bs=384k conv=notrunc
+        [ "$?" -eq "0" ] || fail "Writing u-boot failed"
     fi
 }
 
@@ -187,6 +192,7 @@ function brp_image_drv_redundant_rootfs_w_data()
         umount /var : \
         glob cp-a /* /mnt/root2/ : \
 
+    [ "$?" -eq "0" ] || fail "Creating image failed"
 }
 
 #
